+++ CHANGELOG.md
@@ -6,11 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
-## Version 1.12.5, 7/1/2019
+## Version 1.12.6, 7/2/2019
 
 ### Added
 
-N/A
+1. Periodic routing table integrity check (15 minutes)
+2. Set kafka read pointer to the beginning for new application instances except presence monitor
 
 ### Removed